+++ frontend/src/i18n/utils.ts
@@ -5,8 +5,8 @@ export type { SupportedLocale } from './locale-registry'
 
 import type { SupportedLocale } from './locale-registry'
 
-export const DEFAULT_LOCALE: SupportedLocale = 'zh'
-export const FALLBACK_LOCALE: SupportedLocale = 'zh'
+export const DEFAULT_LOCALE: SupportedLocale = 'vi'
+export const FALLBACK_LOCALE: SupportedLocale = 'vi'
 export const PREFERRED_LOCALE_STORAGE_KEY = 'preferredLocale'
 export const EMPTY_LOCALE_MESSAGES = Object.fromEntries(
   SUPPORTED_LOCALES.map((supportedLocale) => [supportedLocale, {}]),